twitterclone

+++ app/Http/Controllers/TasksController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Task;
+use App\User;
 
 class TasksController extends Controller
 {
@@ -16,11 +17,23 @@ class TasksController extends Controller
     public function index()
     {
     
-        $tasks = Task::all();
-     
-         return view('tasks.index',[
+        $data = [];
+        if (\Auth::check()) {
+        
+        $user = \Auth::user();
+        $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
+        
+        $data = [
+                'user' => $user,
+                            ];
+          return view('tasks.index',[
            'tasks' => $tasks ,
-         ]);
+         ]);                           
+             
+        }
+        return view('welcome', $data);                    
+    
+
     
     }
 
@@ -50,17 +63,18 @@ class TasksController extends Controller
     {
     
         $request->validate([
+            'content' => 'required|max:255',
             'status' => 'required|max:10',
         ]);
     
     
-    
-        $task = new Task;
-        $task->content = $request->content;
-        $task->status = $request->status;
-        $task->save();
+         $request->user()->tasks()->create([
+            'content' => $request->content,
+            'status' => $request->status,
+        ]);
         
-        return redirect('/');
+
+        return back();
         
     }
 
@@ -72,11 +86,16 @@ class TasksController extends Controller
      */
     public function show($id)
     {
-        $task = Task::findOrFail($id);
         
-        return view('tasks.show',[
-            'task'=> $task,
-            ]);
+            $task = Task::findOrFail($id);
+        if (\Auth::id() === $task->user_id) {
+            return view('tasks.show',[
+                 'task'=> $task,
+                ]);
+            
+            
+        }
+
     }
 
     /**
@@ -130,10 +149,12 @@ class TasksController extends Controller
     public function destroy($id)
     {
         $task = Task::findOrFail($id);
-        $task->delete();
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+        }
 
         
-        return redirect('/');
+        return back();
 
     }
 }
